Removing the handler
